refactor: moved 'fromRateToValue' and 'fromValueToRate' to utils.
