+++ lib/utils.ts
@@ -1,0 +1,19 @@
+import { BigNumber } from "ethers";
+
+export function fromRateToValue(
+  contributionRate: BigNumber,
+  perSecondFeeNumerator: BigNumber,
+  perSecondFeeDenominator: BigNumber
+) {
+  return contributionRate
+    .mul(perSecondFeeDenominator)
+    .div(perSecondFeeNumerator);
+}
+
+export function fromValueToRate(
+  value: BigNumber,
+  perSecondFeeNumerator: BigNumber,
+  perSecondFeeDenominator: BigNumber
+) {
+  return value.mul(perSecondFeeNumerator).div(perSecondFeeDenominator);
+}
